Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/css/css-transitions/parsing/transition-duration-computed-expected.txt


PASS Property transition-duration value '500ms' computes to '0.5s' 
PASS Property transition-duration value 'calc(2 * 3s)' computes to '6s' 
FAIL Property transition-duration value '20s, 10s' computes to '20s, 10s' assert_equals: expected "20s, 10s" but got "10s"
